requirements.txt
anthropic>=0.40.0
torch>=2.0.0
transformers>=4.40.0
sentence-transformers>=2.6.0
spacy>=3.7.0
nltk>=3.8.0
numpy>=1.26.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
loguru>=0.7.2
pytest>=8.0.0
pytest-asyncio>=0.23.0
rich>=13.7.0
click>=8.1.7
tiktoken>=0.6.0
regex>=2024.1.0
pdfminer.six>=20221105
python-docx>=1.1.0
